Add start method to ds_reader to keep reader thread from starting early
This was making unit tests impossible as the reader thread would start before the test framework, even in tests where we don't want the reader thread at all.  There are some architecture changes we should make in the future to better separate functionality.

TODO: Add corresponding start method to JNI and call it